fix: add login validation to question routes
This commit adds login validation to PUT and DELETE method routes for question, so that it can compare the ID of the logged in user with the user_id attribute of the requested question.

+++ routes/question.js
@@ -1,5 +1,5 @@
-var express = require("express");
-var router = express.Router();
+const express = require("express");
+const router = express.Router();
 const {
   Question,
   Answer,
@@ -12,6 +12,7 @@ const {
   Bookmark,
   TestQuestion,
 } = require("../models");
+const { isLoggedIn, getLoggedInUserId } = require("./middlewares");
 
 // 전체 문제 리스트를 불러옴
 router.get("/all", async (req, res, next) => {
@@ -48,7 +49,7 @@ router.get("/all", async (req, res, next) => {
 router.get("/:id", async (req, res, next) => {
   try {
     const question = await Question.findOne({
-      where: { id },
+      where: { id: req.params.id },
     });
     return res.status(200).json({
       success: true,
@@ -64,32 +65,39 @@ router.get("/:id", async (req, res, next) => {
 });
 
 // 문제 생성
-router.post("/", async (req, res, next) => {
-  const { content, course_id, user_id } = req.body;
+router.post("/", isLoggedIn, async (req, res, next) => {
+  const { content, course_id } = req.body;
   try {
     // 동일하거나 유사한 문제가 있는 경우
     // TODO: 동일하거나 유사한 문제 존재시 중복문제 또는 복수정답 처리
     // ...
 
+    // 접속한 사용자의 id를 받아옴
+    const user_id = await getLoggedInUserId(req, res);
+
     // 문제 생성
+    const question = await Question.create({
+      content,
+      course_id,
+      user_id,
+    });
+
+    // 생성한 문제에 댓글 및 좋아요 entity id 연결
     const commentable_entity = await CommentableEntity.create({
       entity_type: "question",
     });
     const likeable_entity = await LikeableEntity.create({
       entity_type: "question",
     });
-
-    const question = await Question.create({
-      content,
-      course_id,
-      user_id,
-      commentable_entity_id: commentable_entity.id,
-      likeable_entity_id: likeable_entity.id,
-    });
+    await Question.update(
+      { commentable_entity_id: commentable_entity.id, likeable_entity_id: likeable_entity.id },
+      { where: { id: question.id } }
+    );
 
     return res.json({
       success: true,
       message: "문제가 성공적으로 등록되었습니다.",
+      question: { id: question.id, content, course_id, user_id },
     });
   } catch (error) {
     console.error(error);
@@ -101,8 +109,20 @@ router.post("/", async (req, res, next) => {
 });
 
 // 문제 id에 해당하는 문제 내용을 수정
-router.put("/:id", async (req, res, next) => {
+router.put("/:id", isLoggedIn, async (req, res, next) => {
   try {
+    // 접속한 사용자의 id를 받아옴
+    const user_id = await getLoggedInUserId(req, res);
+
+    // 접속한 사용자의 id와 query에 있는 문제의 user_id룰 를 대조
+    const question = await Question.findOne({ where: { id: req.params.id } });
+    if (question.user_id !== user_id) {
+      return res.status(400).json({
+        success: false,
+        message: "자신이 업로드한 문제만 내용을 수정할 수 있습니다",
+      });
+    }
+
     await Question.update({ content: req.body.content }, { where: { id: req.params.id } });
     return res.status(200).json({
       success: true,
@@ -118,10 +138,19 @@ router.put("/:id", async (req, res, next) => {
 });
 
 // 문제 id에 해당하는 문제 완전 삭제
-router.delete("/:id", async (req, res, next) => {
+router.delete("/:id", isLoggedIn, async (req, res, next) => {
   try {
-    // 문제가 존재하는지 확인
+    // 접속한 사용자의 id를 받아옴
+    const user_id = await getLoggedInUserId(req, res);
+
+    // 접속한 사용자의 id와 query에 있는 문제의 user_id룰 를 대조
     const question = await Question.findOne({ where: { id: req.params.id }, raw: true });
+    if (question.user_id !== user_id) {
+      return res.status(400).json({
+        success: false,
+        message: "자신이 업로드한 문제만 삭제할 수 있습니다",
+      });
+    }
 
     // 문제 삭제
     const result = await Question.destroy({ where: { id: req.params.id } });
